change long to int

+++ projGenericService/src/main/java/ua/tqs/deliveryservice/model/Rider.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Rider extends User {
 
     private long reviewsSum;
-    private long totalNumReviews;
+    private int totalNumReviews;
 
     @OneToMany
     private List<Order> orders;
